# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/HealthResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Liveness + readiness: 200 when the API and its database are reachable, 503 otherwise.",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/products": {
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search over name, description, brand and category.",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 60,
              "type": "string"
            }
          },
          {
            "name": "minPriceCents",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPriceCents",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "description": "Only return products with stock > 0.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Search and filter the active catalog.",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/categories": {
      "get": {
        "operationId": "ProductsController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Distinct categories of active products.",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "A single active product.",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/admin/products": {
      "post": {
        "operationId": "AdminProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "Creates a product.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/products/{id}": {
      "patch": {
        "operationId": "AdminProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "Partial update. Stock is adjusted only via a relative `stockDelta`.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/orders": {
      "post": {
        "operationId": "OrdersController_place",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Places an order. Prices and totals are resolved server-side.",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OrderResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The caller's orders, newest first.",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One of the caller's orders (404 for anyone else's).",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "CustomersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/CustomerResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error envelope (4xx/5xx)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The authenticated customer.",
        "tags": [
          "customers"
        ]
      }
    }
  },
  "info": {
    "title": "CommercePilot API",
    "description": "E-commerce backend with a conversational assistant. Every response uses the envelope `{ success, data, meta }` or `{ success: false, error: { code, message, details?, requestId } }`.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key"
      }
    },
    "schemas": {
      "DependencyCheckDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "up"
            ]
          },
          "latencyMs": {
            "type": "number"
          }
        },
        "required": [
          "status",
          "latencyMs"
        ]
      },
      "HealthChecksDto": {
        "type": "object",
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DependencyCheckDto"
          }
        },
        "required": [
          "database"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ]
          },
          "version": {
            "type": "string"
          },
          "uptimeSeconds": {
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "checks": {
            "$ref": "#/components/schemas/HealthChecksDto"
          }
        },
        "required": [
          "status",
          "version",
          "uptimeSeconds",
          "timestamp",
          "checks"
        ]
      },
      "MetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "4f1c2b8e-1d2a-4c7e-9a51-2f0f1f7f1a10"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "4f1c2b8e-1d2a-4c7e-9a51-2f0f1f7f1a10"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 48
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "requestId",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "items.0.quantity"
          },
          "messages": {
            "example": [
              "quantity must not be less than 1"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "field",
          "messages"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "requestId": {
            "type": "string",
            "example": "4f1c2b8e-1d2a-4c7e-9a51-2f0f1f7f1a10"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "MoneyDto": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "number",
            "example": 19900
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "formatted": {
            "type": "string",
            "example": "$199.00"
          }
        },
        "required": [
          "amountCents",
          "currency",
          "formatted"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "stock": {
            "type": "number"
          },
          "inStock": {
            "type": "boolean",
            "description": "Derived convenience flag: `stock > 0`."
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "description",
          "category",
          "brand",
          "price",
          "stock",
          "inStock",
          "attributes",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9._-]+$"
          },
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "maxLength": 60
          },
          "brand": {
            "type": "string",
            "maxLength": 60
          },
          "priceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "stock": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "sku",
          "name",
          "category",
          "priceCents",
          "stock"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9._-]+$"
          },
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "maxLength": 60
          },
          "brand": {
            "type": "string",
            "maxLength": 60
          },
          "priceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "description": "Price snapshot at purchase time.",
            "allOf": [
              {
                "$ref": "#/components/schemas/MoneyDto"
              }
            ]
          },
          "lineTotal": {
            "$ref": "#/components/schemas/MoneyDto"
          }
        },
        "required": [
          "productId",
          "sku",
          "name",
          "quantity",
          "unitPrice",
          "lineTotal"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "PENDING",
              "CONFIRMED",
              "SHIPPED",
              "CANCELLED"
            ],
            "type": "string",
            "example": "CONFIRMED"
          },
          "id": {
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "itemCount": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "total",
          "itemCount",
          "items",
          "createdAt"
        ]
      },
      "OrderLineDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 50
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CustomerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      }
    }
  }
}